requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.7
django-countries==5.3.1
django-country-and-language-flags==0.1.1
django-debug-toolbar==1.9.1
django-js-asset==1.1.0
django-map-widgets==0.1.9
django-modeltranslation==0.12.2
django-mptt==0.9.0
django-rosetta==0.8.1
django-tables2==1.21.2
idna==2.7
microsofttranslator==0.8
packaging==17.1
Pillow==5.2.0
pluggy==0.6.0
polib==1.1.0
py==1.5.4
py-translate==1.0.3
pyparsing==2.2.0
pytz==2018.5
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
tox==3.1.2
transliterate==1.10.1
urllib3==1.23
virtualenv==16.0.0
